+++ packages/template-retail-react-app/app/utils/responsive-image.js
@@ -18,32 +18,56 @@ const getBreakpointLabels = (breakpoints) =>
         .map(([key]) => key)
 
 const {breakpoints: defaultBreakpoints} = theme
+const defaultDensityFactors = [
+    [1, 1],
+    [2, 2]
+]
+const getEffectiveDensities = (breakpoints) =>
+    Object.entries(breakpoints)
+        .sort((a, b) => parseFloat(a[1]) - parseFloat(b[1]))
+        .map(() => defaultDensityFactors)
+
 let themeBreakpoints = defaultBreakpoints
 let breakpointLabels = getBreakpointLabels(themeBreakpoints)
+let effectiveDefaultDensities = getEffectiveDensities(defaultBreakpoints)
+
+// Init densities cache
+const densitiesCache = new WeakMap()
+densitiesCache.set(themeBreakpoints, new WeakMap())
 
 /**
  * @param {Object} props
  * @param {string} props.src - Dynamic src having an optional param that can vary with widths. For example: `image[_{width}].jpg` or `image.jpg[?sw={width}&q=60]`
  * @param {(number[]|string[]|Object)} [props.widths] - Image widths relative to the breakpoints, whose units can either be px or vw or unit-less. They will be mapped to the corresponding `sizes` and `srcSet`.
+ * @param {(number[]|[number, number][])} [props.densities] - Image density factors to apply relative to the breakpoints. Will be mapped to the corresponding `srcSet`.
  * @param {Object} [props.breakpoints] - The current theme's breakpoints. If not given, Chakra's default breakpoints will be used.
  * @return {Object} src, sizes, and srcSet props for your image component
  */
-export const getResponsiveImageAttributes = ({src, widths, breakpoints = defaultBreakpoints}) => {
+export const getResponsiveImageAttributes = ({
+    src,
+    widths,
+    densities,
+    breakpoints = defaultBreakpoints
+}) => {
     if (!widths) {
         return {
             src: getSrcWithoutOptionalParams(src)
         }
     }
 
-    themeBreakpoints = breakpoints
-    breakpointLabels = getBreakpointLabels(themeBreakpoints)
+    if (breakpoints !== themeBreakpoints) {
+        !densitiesCache.has(breakpoints) && densitiesCache.set(breakpoints, new WeakMap())
+        themeBreakpoints = breakpoints
+        breakpointLabels = getBreakpointLabels(themeBreakpoints)
+        effectiveDefaultDensities = getEffectiveDensities(themeBreakpoints)
+    }
 
     // Order of these attributes matter! If src is not last, Safari will refetch images
     // multiple times (once when it processes src and again when it processes sizes / srcSet)
     // See https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2223
     return {
         sizes: mapWidthsToSizes(widths),
-        srcSet: mapWidthsToSrcSet(widths, src),
+        srcSet: mapWidthsToSrcSet({src, widths, densities}),
         src: getSrcWithoutOptionalParams(src)
     }
 }
@@ -53,7 +77,7 @@ export const getResponsiveImageAttributes = ({src, widths, breakpoints = default
  * @return {string}
  */
 const mapWidthsToSizes = (widths) => {
-    const _widths = withUnit(Array.isArray(widths) ? widths : widthsAsArray(widths))
+    const _widths = withUnit(Array.isArray(widths) ? widths : breakpointMapAsArray(widths))
 
     return breakpointLabels
         .slice(0, _widths.length)
@@ -64,29 +88,90 @@ const mapWidthsToSizes = (widths) => {
         .join(', ')
 }
 
+function padArray(arr) {
+    const l1 = arr.length
+    const l2 = breakpointLabels.length
+    if (l1 < l2) {
+        const lastEntry = arr.at(-1)
+        const amountToPad = l2 - l1
+        return [...arr, ...Array(amountToPad).fill(lastEntry)]
+    }
+    return arr
+}
+
 /**
- * @param {(number[]|string[]|Object)} widths
+ * @param {(number[]|[number, number][])} densities
+ * @param {number[][][]} [collector]
+ * @returns {number[][][]}
+ */
+function mapDensities(densities, collector) {
+    const flat = Array.isArray(collector)
+    return densities.reduce(
+        (acc, entry) => {
+            if (!flat && Array.isArray(entry) && Array.isArray(entry.at(0))) {
+                acc.push(mapDensities(entry, []))
+            } else if (
+                Array.isArray(entry) &&
+                typeof entry.at(0) === 'number' &&
+                !Number.isNaN(entry.at(0))
+            ) {
+                const w = entry.at(0)
+                const f = entry.at(1)
+                const result = [w, typeof f === 'number' && !Number.isNaN(f) ? f : w]
+                acc.push(flat ? result : [result])
+            } else if (typeof entry === 'number' && !Number.isNaN(entry)) {
+                const result = [entry, entry]
+                acc.push(flat ? result : [result])
+            } else if (flat) {
+                acc.push(...defaultDensityFactors)
+            } else {
+                acc.push(defaultDensityFactors)
+            }
+            return acc
+        },
+        flat ? collector : []
+    )
+}
+
+function obtainDensities(densities) {
+    const cache = densitiesCache.get(themeBreakpoints)
+    if (cache?.has(densities)) {
+        return densitiesCache.get(themeBreakpoints).get(densities)
+    } else if (isObject(densities)) {
+        const _densities = mapDensities(padArray(breakpointMapAsArray(densities)))
+        cache?.set(densities, _densities)
+        return _densities
+    } else if (Array.isArray(densities)) {
+        const _densities = mapDensities(padArray(densities))
+        cache?.set(densities, _densities)
+        return _densities
+    }
+    return effectiveDefaultDensities
+}
+
+/**
+ * @param {Object} props
+ * @param {string} props.src
+ * @param {(number[]|string[]|Object)} props.widths
+ * @param {(number[]|[number, number][])} props.densities
  * @return {string}
  */
-const mapWidthsToSrcSet = (widths, dynamicSrc) => {
-    let _widths = isObject(widths) ? widthsAsArray(widths) : widths.slice(0)
+const mapWidthsToSrcSet = ({src, widths, densities}) => {
+    let _widths = isObject(widths) ? breakpointMapAsArray(widths) : widths.slice(0)
+    _widths = uniqueArray(convertToPxNumbers(padArray(_widths))).sort()
 
-    if (_widths.length < breakpointLabels.length) {
-        const lastWidth = _widths[_widths.length - 1]
-        const amountToPad = breakpointLabels.length - _widths.length
-
-        _widths = [..._widths, ...Array(amountToPad).fill(lastWidth)]
-    }
-
-    _widths = uniqueArray(convertToPxNumbers(_widths)).sort()
-
-    const srcSet = []
-    _widths.forEach((width) => {
-        srcSet.push(width)
-        srcSet.push(width * 2) // for devices with higher pixel density
-    })
-
-    return srcSet.map((imageWidth) => `${getSrc(dynamicSrc, imageWidth)} ${imageWidth}w`).join(', ')
+    // Request images using given or default density factors. By default, images are requested with factors 1 and 2.
+    const _densities = obtainDensities(densities)
+    const set = _widths.reduce((acc, width, idx) => {
+        const densityFactors = _densities.at(idx)
+        for (const [w, f] of densityFactors) {
+            const w1 = Math.round(width * w)
+            const w2 = Math.round(width * f)
+            acc.add(`${getSrc(src, w2)} ${w1}w`)
+        }
+        return acc
+    }, new Set())
+    return [...set].join(', ')
 }
 
 const vwValue = /^\d+vw$/
@@ -140,22 +225,21 @@ const withUnit = (widths) =>
 const isObject = (o) => o?.constructor === Object
 
 /**
- * @param {Object} widths
+ * @param {Object} map
  * @example
  * // returns the array [10, 10, 10, 50]
- * widthsAsArray({base: 10, lg: 50})
+ * breakpointMapAsArray({base: 10, lg: 50})
  */
-const widthsAsArray = (widths) => {
-    const biggestBreakpoint = breakpointLabels.filter((bp) => Boolean(widths[bp])).pop()
+const breakpointMapAsArray = (map) => {
+    const biggestBreakpoint = breakpointLabels.filter((bp) => Boolean(map[bp])).pop()
 
     let mostRecent
     return breakpointLabels.slice(0, breakpointLabels.indexOf(biggestBreakpoint) + 1).map((bp) => {
-        if (widths[bp]) {
-            mostRecent = widths[bp]
-            return widths[bp]
-        } else {
-            return mostRecent
+        if (map[bp]) {
+            mostRecent = map[bp]
+            return map[bp]
         }
+        return mostRecent
     })
 }